allow empty string for some param fields

+++ bin/rpt_transfer.py
@@ -3,17 +3,11 @@ import shutil
 import logging
 from rpt_config import validate_path
 
-def transfertC4(config):
-    validate_path(config["source_path"].get_value(), "source_path")
-    validate_path(config["to_clean"].get_value(), "to_clean")
-    validate_path(config["dest"].get_value(), "dest")
-    validate_path(config["dest_JBE"].get_value(), "dest_JBE")
-
-
-    source_path = config["source_path"].get_value()
-    to_clean_path = config["to_clean"].get_value()
-    dest_path = config["dest"].get_value()
-    dest_JBE_path = config["dest_JBE"].get_value()
+def transfer(config):
+    source_path = validate_path(config["source_path"].get_value(), "source_path")
+    to_clean_path = validate_path(config["to_clean"].get_value(), "to_clean", none_allowed=True)
+    dest_path = validate_path(config["dest"].get_value(), "dest")
+    dest_JBE_path = validate_path(config["dest_JBE"].get_value(), "dest_JBE")
 
     if os.name == "nt":
         files = [f for f in os.listdir(source_path) if (os.path.isfile(os.path.join(source_path,f)) and os.path.join(source_path,f).endswith(".pdf") and os.path.join(source_path,f).startswith("C"))]
@@ -45,12 +39,13 @@ def transfertC4(config):
         logging.warning(f"{len(not_transfered)} fichiers non transférés (déjà présents dans {dest_path}):")
         for f in not_transfered:
             logging.warning(f"{f} non transféré")
-    
-    to_del = [os.path.join(to_clean_path,f) for f in os.listdir(to_clean_path) if os.path.isfile(os.path.join(to_clean_path,f)) ]
-    logging.info(f"{len(to_del)} fichiers à supprimer de {to_clean_path}")
-    for file in to_del:
-        logging.info(f"Suppression de {file}")
-        os.remove(file)
+
+    if to_clean_path is not None:    
+        to_del = [os.path.join(to_clean_path,f) for f in os.listdir(to_clean_path) if os.path.isfile(os.path.join(to_clean_path,f)) ]
+        logging.info(f"{len(to_del)} fichiers à supprimer de {to_clean_path}")
+        for file in to_del:
+            logging.info(f"Suppression de {file}")
+            os.remove(file)
 
     if config.get("open_explorer") and config["open_explorer"].get_value():
         logging.info(f"Ouverture de {source_path} dans l'explorateur de fichier")
